% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htx_load.R
\name{addRD}
\alias{addRD}
\title{add gene-level rowData derived from transcript level rowRanges}
\usage{
addRD(x)
}
\arguments{
\item{x}{result of htx_load()}
}
\value{
RangedSummarizedExperiment with enhanced rowData
}
\description{
add gene-level rowData derived from transcript level rowRanges
}
\examples{
# this function operates on a SummarizedExperiment that has
# transcript-level rowRanges but gene-level quantifications
addRD
}
