% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca43kAnVIL.R
\name{ca43kAnVIL}
\alias{ca43kAnVIL}
\title{app to survey 43000 cancer transcriptomes}
\usage{
ca43kAnVIL()
}
\value{
a SummarizedExperiment
}
\description{
app to survey 43000 cancer transcriptomes
}
\note{
Copies source code and metadata to a temporary folder
and executes shiny::runApp there; sets working diretory to folder where
ca43kAnVIL was called
when app is exited.  Also will return either NULL or a
SummarizedExperiment at conclusion.
}
\examples{
ca43kAnVIL
}
