% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ds4842}
\alias{ds4842}
\title{ds4842: ssrch::DocSet instance with most BigRNA studies}
\format{DocSet class instance}
\usage{
ds4842
}
\description{
ds4842: ssrch::DocSet instance with most BigRNA studies
}
\keyword{datasets}
