% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{experTable}
\alias{experTable}
\title{experiment-level metadata}
\format{data.frame}
\source{
SRAdbV2 28 June 2018
}
\usage{
experTable
}
\description{
experiment-level metadata
}
\examples{
head(HumanTranscriptomeCompendium::experTable)
}
\keyword{datasets}
