% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPMIDs.R
\name{getPMID}
\alias{getPMID}
\title{use SRAdbV2 to get one PMID for a study}
\usage{
getPMID(acc)
}
\arguments{
\item{acc}{character(1) study accession number}
}
\value{
character string of PMID or NA_character_
}
\description{
use SRAdbV2 to get one PMID for a study
}
\examples{
if (interactive()) getPMID("SRP057500")
}
