% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htx_app.R
\name{htx_app}
\alias{htx_app}
\title{explore SRA metadata}
\usage{
htx_app()
}
\description{
explore SRA metadata
}
\note{
This function deals with
extraction of compendium elements.  The overall scope is
determined by HumanTranscriptomeCompendium::studTable which is the list of
all studies with taxon 9606, strategy RNA-seq,
source transcriptomic.  Some studies will not have
experiments in the compendium, and if such are selected,
a warning will be generated in the session.
}
\examples{
if (interactive()) htx_app()
}
