% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htx_query.R
\name{htx_query}
\alias{htx_query}
\title{retrieve 'restfulSE' SummarizedExperiment instance for selected studies in htx compendium}
\usage{
htx_query(study_accessions, ...)
}
\arguments{
\item{study_accessions}{character vector of study accessions}

\item{\dots}{passed to loadHtxcomp}
}
\value{
SummarizedExperiment instance
}
\description{
retrieve 'restfulSE' SummarizedExperiment instance for selected studies in htx compendium
}
\examples{
htx_query("ERP011411")
}
