% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htx_query.R
\name{htx_query_by_study_accession}
\alias{htx_query_by_study_accession}
\title{retrieve 'restfulSE' SummarizedExperiment instance for selected studies in htx compendium}
\usage{
htx_query_by_study_accession(studies, htxSE, ...)
}
\arguments{
\item{studies}{character vector of study accessions}

\item{htxSE}{SummarizedExperiment instance, typically the result of htx_load(), which we don't want to repeat needlessly}

\item{\dots}{passed to `htx_load`, ignored if `se` is nonmissing}
}
\value{
SummarizedExperiment instance
}
\description{
retrieve 'restfulSE' SummarizedExperiment instance for selected studies in htx compendium
}
\note{
This function was designed to perform a single
query on a 'fresh' compendium image from `htx_load()`.  However,
one could consider iterating the process to build up
metadata on multiple series of studies.  This is not likely
to succeed without careful manipulation of the colData of the
input SummarizedExperiment.  A message will be written if
the input SummarizedExperiment appears to be other than a 'fresh'
`htx_load` result.
}
\examples{
htx_query_by_study_accession("ERP011411")
}
