% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htx_query.R
\name{htx_query_by_text}
\alias{htx_query_by_text}
\title{subset compendium through keyword lookup}
\usage{
htx_query_by_text(query, ...)
}
\arguments{
\item{query}{character(1) to be found in ls(ssrch::kw2docs(ds4842))}

\item{\dots}{passed to `htx_query_by_study_accession`}
}
\value{
SummarizedExperiment instance
}
\description{
subset compendium through keyword lookup
}
\note{
The DocSet instance ds4842 is used.  Lookups are case-sensitive.
Look carefully at note for `htx_query_by_study_accession` to
understand logic of incrementing metadata on a given
input SummarizedExperiment.
}
\examples{
htx_query_by_text("HNRNPC")
}
