% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HumanTranscriptomeCompendium.colnames}
\alias{HumanTranscriptomeCompendium.colnames}
\title{character vector of available samples in HDF cloud assay representation}
\format{
character vector
}
\source{
compendium processing
}
\usage{
HumanTranscriptomeCompendium.colnames
}
\description{
character vector of available samples in HDF cloud assay representation
}
\examples{
head(HumanTranscriptomeCompendium::HumanTranscriptomeCompendium.colnames)
}
\keyword{datasets}
