% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_works.R
\name{load_bigrnaFiles}
\alias{load_bigrnaFiles}
\title{obtain listing of contents of BigRNA compendium (salmon runs)}
\usage{
load_bigrnaFiles(cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{cache}{instance of `BiocFileCache`, defaults to `BiocFileCache::BiocFileCache()`}
}
\value{
a named vector
}
\description{
obtain listing of contents of BigRNA compendium (salmon runs)
}
\examples{
if (interactive()) head(load_bigrnaFiles())
}
