% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_works.R
\name{load_experTable}
\alias{load_experTable}
\title{obtain listing of experiments and submission date/time in compendium}
\usage{
load_experTable(cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{cache}{instance of `BiocFileCache`, defaults to `BiocFileCache::BiocFileCache()`}
}
\value{
a data.frame
}
\description{
obtain listing of experiments and submission date/time in compendium
}
\examples{
if (interactive()) head(load_experTable())
}
