% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_works.R
\name{load_studTable}
\alias{load_studTable}
\title{obtain listing of all studies in compendium}
\usage{
load_studTable(cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{cache}{instance of `BiocFileCache`, defaults to `BiocFileCache::BiocFileCache()`}
}
\value{
a data.frame
}
\description{
obtain listing of all studies in compendium
}
\examples{
if (interactive()) head(load_studTable())
}
