% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_works.R
\name{path_doc4842}
\alias{path_doc4842}
\title{return path to metadata csvs in zip file}
\usage{
path_doc4842(cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{cache}{instance of `BiocFileCache`, defaults to `BiocFileCache::BiocFileCache()`}
}
\value{
path to zipfile
}
\description{
return path to metadata csvs in zip file
}
\note{
CSVs were retrieved using methods provided at
\url{https://api-omicidx.cancerdatasci.org/sra/1.0/ui/} and zipped together.  Function
will lodge zipfile in `cache` if not present.
}
\examples{
path_doc4842()
}
