% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifTools.R
\name{procExpToGene}
\alias{procExpToGene}
\title{acquire a single sample from bigRNA compendium specified by accession
and develop gene-level quantifications using tximport}
\usage{
procExpToGene(acc, tx2gene = tx2gene_gencode27(),
  urlprefix = "http://bigrna.cancerdatasci.org/",
  manifestdata = HumanTranscriptomeCompendium::bigrnaFiles,
  regexp = "quant.sf.bz2|json")
}
\arguments{
\item{acc}{character(1) sample-level accession as defined in SRA}

\item{tx2gene}{a data.frame instance mapping transcript identifiers
used in the compendium to gene identifiers.  See note.}

\item{urlprefix}{character(1) where the salmon run outputs are lodged,
with \code{acc} a subfolder defined through the manifestData parameter.}

\item{manifestdata}{a character vector defining folders (under
\code{results/human/27/} with salmon outputs.}

\item{regexp}{a character(1) regular expression for filtering
filename elements in \code{manifestdata} to define which salmon output
components in the bigrna compendium are retrieved.}
}
\value{
the result of a tximport run
}
\description{
acquire a single sample from bigRNA compendium specified by accession
and develop gene-level quantifications using tximport
}
\note{
The tx2gene_gencode function supplied with this package
uses the tximportData package contents to create the data.frame
for use as tx2gene.  The system2 function is used to generate
folders to be used by tximport.
}
\examples{
# this example involves nontrivial internet communications
args(procExpToGene)
\dontrun{
td = tempdir()
od = getwd()
setwd(td)
nn = procExpToGene("ERX1097381")
str(nn)
setwd(od)
}
}
