% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{studTable}
\alias{studTable}
\title{study-level metadata}
\format{data.frame}
\source{
SRAdbV2 28 June 2018
}
\usage{
studTable
}
\description{
study-level metadata
}
\examples{
head(HumanTranscriptomeCompendium::studTable)
}
\keyword{datasets}
