% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifTools.R
\name{tx2gene_gencode27}
\alias{tx2gene_gencode27}
\title{generate a data.frame mapping gencode 27 ensembl transcript identifiers to ensembl gene identifiers}
\usage{
tx2gene_gencode27()
}
\value{
a data.frame with 200401 rows mapping transcript 
identifiers in column 1 to 58288 gene symbols in column 2.
}
\description{
generate a data.frame mapping gencode 27 ensembl transcript identifiers to ensembl gene identifiers
}
\note{
Uses CSV in tximportData to acquire the information.
}
\examples{
head(tx2gene_gencode27())
}
