% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uniqueAcc_120518}
\alias{uniqueAcc_120518}
\title{experiment accessions available in compendium as of may 12 2018}
\format{
data.frame
}
\source{
SRAdbV2 may 12 2018
}
\usage{
uniqueAcc_120518
}
\description{
experiment accessions available in compendium as of may 12 2018
}
\examples{
head(HumanTranscriptomeCompendium::uniqueAcc_120518)
}
\keyword{datasets}
