% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htx_query.R
\name{htx_query_by_text}
\alias{htx_query_by_text}
\title{subset compendium through keyword lookup}
\usage{
htx_query_by_text(query, ..., tryGrep = TRUE, ignore.case = TRUE)
}
\arguments{
\item{query}{character(1) to be found in ls(ssrch::kw2docs(get_ds4841()))}

\item{\dots}{passed to `htx_query_by_study_accession`}

\item{tryGrep}{logical(1) if TRUE, `query` does not match any keyword directly, it will be treated as a regular expression and the vector of keywords will be grepped for pattern `query`; defaults to TRUE}

\item{ignore.case}{logical(1) used when tryGrep is TRUE, defaults to TRUE}
}
\value{
SummarizedExperiment instance
}
\description{
subset compendium through keyword lookup
}
\note{
The DocSet instance returned by `get_ds4841()` is used.  Lookups are case-sensitive.
Look carefully at note for `htx_query_by_study_accession` to
understand logic of incrementing metadata on a given
input SummarizedExperiment.
}
\examples{
htx_query_by_text("HNRNPC")
}
